Extract PDF as Pages and TextLines instead of plain text
Poppler's utility `pdftohtml` (one of the Docsplit dependencies) can
extract the PDF content and structure of pages to a friendly XML format.
Firstly, this allow us to perform layout analysis for building a TOC for
reference, for removing useless text chunks that could confuse some part
of Freeling's processing chain, etc.
Secondly, from this XML we can recover how the PDF used to look like
into a HTML version that can be displayed on the browser.  This is an
excellent replacement to the current document viewer, that only shows
the extracted and unstructured plain text (which is a total mess).

Also, we should start migrating "process" logic _away_ from the models
to their Resque tasks classes. It is easier if we do not to mix
"processing" and "persistance" (Mongo).  We may change how we do the
first or the second independently in the near future.